Entregas - Lucas Lima Siade/Projeto Integrador/requirements.txt
google-adk
google-genai
google-generativeai
python-dotenv
feedparser
fpdf